+++ www/wp-content/themes/luziapi/src/Newsletter/Application/Query/GetSubscribers/GetSubscribersQuery.php
@@ -1,0 +1,13 @@
+<?php
+
+declare(strict_types=1);
+
+namespace LuziApi\Newsletter\Application\Query\GetSubscribers;
+
+final readonly class GetSubscribersQuery
+{
+    public function __construct(
+        public string $search = '',
+    ) {
+    }
+}